platform/crm-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
prometheus-client==0.20.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

alembic==1.13.2
python-jose==3.3.0
